新增錯誤處理，所有輸入的檢查(應該)

+++ worker/src/lib.rs
@@ -14,7 +14,7 @@ pub async fn list_all_worker_name() -> String {
     )
 }
 
-fn get_all_worker_names() -> Vec<String> {
+pub fn get_all_worker_names() -> Vec<String> {
     let mut conn = get_conn();
     let query = "SELECT name FROM name_table";
     let names: Vec<String> = conn.query_map(query, |name| name).unwrap();
@@ -28,15 +28,15 @@ pub async fn add_worker_name(req: &mut Request) -> String {
         .unwrap_or("NULL!!".to_string());
 
     if worker_name == "NULL!!" {
-        return "worker_name not provided".to_string();
+        return "Err worker_name not provided".to_string();
     }
 
     let mut conn = get_conn();
 
     let query = "INSERT INTO name_table (name) VALUES (:name)";
     match conn.exec_drop(query, params! { "name" => worker_name }) {
-        Ok(_) => "name insert successful".to_string(),
-        Err(_) => "name insert failed".to_string(),
+        Ok(_) => "Ok".to_string(),
+        Err(_) => "Err".to_string(),
     }
 }
 #[handler]
@@ -46,14 +46,18 @@ pub async fn remove_worker_name(req: &mut Request) -> String {
         .unwrap_or("NULL!!".to_string());
 
     if worker_name == "NULL!!" {
-        return "worker_name not provided".to_string();
+        return "Err worker_name not provided".to_string();
+    }
+
+    if !get_all_worker_names().contains(&worker_name) {
+        return "Err worker_name not exist".to_string();
     }
 
     let mut conn = get_conn();
     //DELETE FROM `todo_list`.`name_table` WHERE (`name` = 'Sandy');
     let query = "DELETE FROM name_table WHERE  (`name` =:name)";
     match conn.exec_drop(query, params! { "name" => worker_name }) {
-        Ok(_) => "worker_name remove successful".to_string(),
-        Err(_) => "worker_name remove failed".to_string(),
+        Ok(_) => "Ok".to_string(),
+        Err(_) => "Err".to_string(),
     }
 }
